chore: move polkadot functions to utils folder
