+++ apps/tangle-dapp/constants/index.ts
@@ -1,2 +1,1 @@
 export * from './evm';
-export * from './polkadot';
