+++ apps/tangle-dapp/data/NominatorStats/useUnbondingRemainingErasSubscription.ts
@@ -5,8 +5,8 @@ import { WebbError, WebbErrorCodes } from '@webb-tools/dapp-types/WebbError';
 import { useEffect, useState } from 'react';
 import { firstValueFrom, type Subscription } from 'rxjs';
 
-import { formatTokenBalance, getPolkadotApiRx } from '../../constants';
 import useFormatReturnType from '../../hooks/useFormatReturnType';
+import { formatTokenBalance, getPolkadotApiRx } from '../../utils/polkadot';
 
 type UnbondingRemainingEras = {
   amount: string;
